\name{gsrmtp-package}
\alias{gsrmtp-package}
\alias{gsrmtp}
\docType{package}
\title{
A graphical approach to sequentially rejective multiple test procedures
}
\description{
This package provides functions and a graphical user interface for graphical described sequentially rejective multiple test procedures.
}
\details{
\tabular{ll}{
Package: \tab gsrmtp\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-02-23\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
%%~~ An overview of how to use the package, including the most important ~~
%%~~ functions ~~
}
\author{
Kornelius Rohmeyer 

Maintainer: Kornelius Rohmeyer <rohmeyer@small-projects.de>
}
\references{
Frank Bretz, Willi Maurer, Werner Brannath, Martin Posch:
A graphical approach to sequentially rejective multiple test procedures.
Statistics in Medicine 2009 vol. 28 issue 4 page 586-604. 
http://www.meduniwien.ac.at/fwf_adaptive/papers/bretz_2009_22.pdf
}
\keyword{ package }
\keyword{ htest }
\seealso{
\code{\link[mutoss:Mutoss]{Mutoss}}
}
\examples{
srmtp()
}
